fix(authentication)removed logindialog.js and created actions.any
