+++ react/features/authentication/components/native/LoginDialog.js
@@ -18,7 +18,8 @@ import {
 import { translate } from '../../../base/i18n';
 import { JitsiConnectionErrors } from '../../../base/lib-jitsi-meet';
 import type { StyleType } from '../../../base/styles';
-import { authenticateAndUpgradeRole, cancelLogin } from '../../actions.native';
+import { authenticateAndUpgradeRole } from '../../actions.any';
+import { cancelLogin } from '../../actions.native';
 
 // Register styles.
 import './styles';
